requirements.txt
black
build
elastic-pipes-core @ git+https://github.com/elastic/pipes-core-py
httpx
isort
ruff
